+++ NEWS.md
@@ -5,6 +5,11 @@ Deprecated the implicit attributes:
 - `Builder.legacy_attributes`
 - `Builder.legacy_long_methods`
 
+## Package API v2.3
+- `find_headers()`: new `strategy` parameter
+- `find_headers()`: new `max_depth` parameter
+- `find_libraries()`: new `strategy` parameter
+
 ## Package API v2.2
 Added to `spack.package`:
 - `BuilderWithDefaults`